Separated out client side commands.
