+++ packages/app/client/src/commands/uiCommands.ts
@@ -1,0 +1,86 @@
+//
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license.
+//
+// Microsoft Bot Framework: http://botframework.com
+//
+// Bot Framework Emulator Github:
+// https://github.com/Microsoft/BotFramwork-Emulator
+//
+// Copyright (c) Microsoft Corporation
+// All rights reserved.
+//
+// MIT License:
+// Permission is hereby granted, free of charge, to any person obtaining
+// a copy of this software and associated documentation files (the
+// "Software"), to deal in the Software without restriction, including
+// without limitation the rights to use, copy, modify, merge, publish,
+// distribute, sublicense, and/or sell copies of the Software, and to
+// permit persons to whom the Software is furnished to do so, subject to
+// the following conditions:
+//
+// The above copyright notice and this permission notice shall be
+// included in all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED ""AS IS"", WITHOUT WARRANTY OF ANY KIND,
+// EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
+// MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+// NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
+// LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
+// WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import { CommandRegistry } from './commandRegistry';
+import { showWelcomePage } from '../data/editorHelpers';
+import { BotCreationDialog, DialogService, SecretPromptDialog } from '../ui/dialogs';
+import store from '../data/store';
+import * as EditorActions from '../data/action/editorActions';
+import * as NavBarActions from '../data/action/navBarActions';
+import * as Constants from '../constants';
+
+/** Register UI commands (toggling UI) */
+export const registerCommands = () => {
+  // ---------------------------------------------------------------------------
+  // Shows the welcome page
+  CommandRegistry.registerCommand('welcome-page:show', () => {
+    showWelcomePage();
+  });
+
+  // ---------------------------------------------------------------------------
+  // Shows a bot creation dialog
+  CommandRegistry.registerCommand('bot-creation:show', () => {
+    DialogService.showDialog(BotCreationDialog);
+  });
+
+  // ---------------------------------------------------------------------------
+  // Shows a dialog prompting the user for a bot secret
+  CommandRegistry.registerCommand('secret-prompt:show', async () => {
+    return await DialogService.showDialog(SecretPromptDialog);
+  });
+
+  // ---------------------------------------------------------------------------
+  // Switches navbar tab selection
+  CommandRegistry.registerCommand('navbar:switchtab', (tabName: string): void => {
+    store.dispatch(NavBarActions.select(tabName));
+  });
+
+  // ---------------------------------------------------------------------------
+  // Switches navbar tab selection to Explorer
+  CommandRegistry.registerCommand('shell:show-explorer', (): void => {
+    store.dispatch(NavBarActions.select(Constants.NAVBAR_BOT_EXPLORER));
+  });
+
+  // ---------------------------------------------------------------------------
+  // Switches navbar tab selection to Services
+  CommandRegistry.registerCommand('shell:show-services', (): void => {
+    store.dispatch(NavBarActions.select(Constants.NAVBAR_SERVICES));
+  });
+
+  // ---------------------------------------------------------------------------
+  // Open App Settings
+  CommandRegistry.registerCommand('shell:show-app-settings', (): void => {
+    const { CONTENT_TYPE_APP_SETTINGS, DOCUMENT_ID_APP_SETTINGS } = Constants;
+    store.dispatch(EditorActions.open(CONTENT_TYPE_APP_SETTINGS, DOCUMENT_ID_APP_SETTINGS, true, null));
+  });
+};
